+++ mink/limits/configuration_limit.py
@@ -4,7 +4,7 @@ import mujoco
 import numpy as np
 
 from ..configuration import Configuration
-from ..constants import qpos_width
+from ..constants import dof_width, qpos_width
 from .exceptions import LimitDefinitionError
 from .limit import Constraint, Limit
 
@@ -39,18 +39,21 @@ class ConfigurationLimit(Limit):
             )
 
         index_list: list[int] = []  # DoF indices that are limited.
-        lower = np.full(model.nq, -np.inf)
-        upper = np.full(model.nq, np.inf)
+        lower = np.full(model.nq, -mujoco.mjMAXVAL)
+        upper = np.full(model.nq, mujoco.mjMAXVAL)
         for jnt in range(model.njnt):
             jnt_type = model.jnt_type[jnt]
             qpos_dim = qpos_width(jnt_type)
             jnt_range = model.jnt_range[jnt]
             padr = model.jnt_qposadr[jnt]
+            # Skip free joints and joints without limits.
             if jnt_type == mujoco.mjtJoint.mjJNT_FREE or not model.jnt_limited[jnt]:
                 continue
             lower[padr : padr + qpos_dim] = jnt_range[0] + min_distance_from_limits
             upper[padr : padr + qpos_dim] = jnt_range[1] - min_distance_from_limits
-            index_list.append(model.jnt_dofadr[jnt])
+            jnt_dim = dof_width(jnt_type)
+            jnt_id = model.jnt_dofadr[jnt]
+            index_list.extend(range(jnt_id, jnt_id + jnt_dim))
 
         self.indices = np.array(index_list)
         self.indices.setflags(write=False)
@@ -89,9 +92,11 @@ class ConfigurationLimit(Limit):
             :math:`G \Delta q \leq h`, or ``None`` if there is no limit.
         """
         del dt  # Unused.
+        if self.projection_matrix is None:
+            return Constraint()
 
         # Upper.
-        delta_q_max = np.zeros(self.model.nv)
+        delta_q_max = np.zeros((self.model.nv,))
         mujoco.mj_differentiatePos(
             m=self.model,
             qvel=delta_q_max,
@@ -101,7 +106,7 @@ class ConfigurationLimit(Limit):
         )
 
         # Lower.
-        delta_q_min = np.zeros(self.model.nv)
+        delta_q_min = np.zeros((self.model.nv,))
         mujoco.mj_differentiatePos(
             m=self.model,
             qvel=delta_q_min,