+++ tests/test_lie_utils.py
@@ -1,4 +1,4 @@
-"""Tests utils.py."""
+"""Tests lie/utils.py."""
 
 import numpy as np
 from absl.testing import absltest